Finance assistant

+++ app/src/main/java/com/localexpense/tracker/data/CategoryTotal.kt
@@ -2,5 +2,5 @@ package com.localexpense.tracker.data
 
 data class CategoryTotal(
     val categoryName: String,
-    val total: Double
+    val total: Long            // إجمالي بوحدات صغرى (قروش)
 )
