+++ app/src/main/java/com/localexpense/tracker/data/SourceTotal.kt
@@ -2,5 +2,5 @@ package com.localexpense.tracker.data
 
 data class SourceTotal(
     val bankName: String,
-    val total: Double
+    val total: Long            // إجمالي بوحدات صغرى (قروش)
 )
